+++ packages/nodes/table/src/components/TableCellElement/useTableCellElementState.ts
@@ -1,9 +1,11 @@
 import { TElement, useElement, usePlateEditorRef } from '@udecode/plate-common';
 import { useReadOnly } from 'slate-react';
+import { ELEMENT_TR } from '../../createTablePlugin';
 import { ELEMENT_TABLE, ELEMENT_TR } from '../../createTablePlugin';
 import { getTableColumnIndex } from '../../queries';
 import { useTableRowStore } from '../../stores/tableRowStore';
 import { useTableStore } from '../../stores/tableStore';
+import { TTableRowElement } from '../../types';
 import { TTableElement } from '../../types';
 import { useIsCellSelected } from './useIsCellSelected';
 
@@ -35,6 +37,7 @@ export const useTableCellElementState = ({
   const rowSizeOverride = useTableRowStore().get.overrideSize();
   const rowSize = rowSizeOverride ?? rowElement?.size ?? undefined;
 
+
   const readOnly = useReadOnly();
 
   const colIndex = getTableColumnIndex(editor, cellElement);
@@ -44,9 +47,9 @@ export const useTableCellElementState = ({
     readOnly: !ignoreReadOnly && readOnly,
     selected: isCellSelected,
     hovered: hoveredColIndex === colIndex,
+    rowSize,
     isLastCell: colIndex === rowElement?.children.length - 1,
     isLastRow:
       tableElement.children[tableElement.children.length - 1] === rowElement,
-    rowSize,
   };
 };