+++ packages/nodes/table/src/transforms/index.ts
@@ -11,3 +11,4 @@ export * from './insertTableRow';
 export * from './moveSelectionFromCell';
 export * from './overrideSelectionFromCell';
 export * from './setTableColSize';
+export * from './setTableRowSize';