+++ packages/nodes/table/src/types.ts
@@ -49,8 +49,18 @@ export interface TablePlugin<V extends Value = Value> {
    * Merged cells not supported.
    */
   initialTableWidth?: number;
+
+  /**
+   * The minimum width of a column.
+   * @default 48
+   */
+  minColumnWidth?: number;
 }
 
 export interface TTableElement extends TElement {
   colSizes?: number[];
+}
+
+export interface TTableRowElement extends TElement {
+  size?: number;
 }