+++ packages/ui/plate/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @udecode/plate-ui
 
+## 20.2.0
+
+## 20.1.0
+
+## 20.0.1
+
 ## 20.0.0
 
 ### Minor Changes